pyproject.toml
[project]
name = "robot-framework-101"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
package-mode = true
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "robotframework (>=7.4.1,<8.0.0)",
    "robotframework-requests (>=0.9.7,<0.10.0)",
    "pymongo (==4.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "robotframework-dashboard (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

